fix: move test to proper file and do not fail on interruption
